requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Werkzeug==3.0.6
flask
Flask-Cors
flask-socketio
python-dotenv
openai
anthropic
google-generativeai
